requirements.txt
matplotlib>=3.8
python-pptx>=1.0
tqdm>=4.66
